Define a packet
Since we're going to need to simulate TCP we need to define the basic fields for
a packet.

The Syn/Ack fields are 64 bits as they don't really take up any bandwidth so
there is little cost to making them large, and we don't want to have to deal
with integer rollover.